fix some build warning.

+++ foundation/unitedb/include/unitedb/table.hpp
@@ -155,7 +155,7 @@ namespace unitedb
 		{
 			// note, when combine with self, do nothing.so "i > 0".
 
-			for (int i = stack_->infos_.size() - 1; i > 0; --i)
+			for (int64_t i = stack_->infos_.size() - 1; i > 0; --i)
 			{
 				if (stack_->infos_[i].rev_ == revision)
 				{
@@ -167,7 +167,7 @@ namespace unitedb
 
 		virtual void Commit(DBRevision revision) override
 		{
-			int idx = -1;
+			int64_t idx = -1;
 			for (int i = 0; i < stack_->infos_.size(); ++i)
 			{
 				if (stack_->infos_[i].rev_ == revision)
@@ -185,12 +185,12 @@ namespace unitedb
 
 			const TableUndoInfo& revinfo = stack_->infos_[idx];
 
-			const int remcount = revinfo.begin_;
+			const int64_t remcount = revinfo.begin_;
 
 			stack_->cache_.Remove(0, remcount);
 
 			// remove info data.
-			int fixcount = idx + 1;
+			int64_t fixcount = idx + 1;
 			auto remend = stack_->infos_.begin() + fixcount;
 
 			stack_->infos_.erase(stack_->infos_.begin(), remend);
